Migrated to Provider for Search Page
