+++ src/components/SearchPage/SearchPage.tsx
@@ -1,19 +1,30 @@
 import * as React from 'react';
 
-import './index.scss';
+import { connect } from 'react-redux';
+
 import SearchFilterPane from '@component/SearchFiltersPane';
 import SearchContent from '@component/SearchContent';
+import { CONST_STORAGE, CONST_URLS } from '@data/Constants';
+
+import './index.scss';
+import { Product } from '@data/Product';
 interface SearchPageProps {
-    searchfilters:{
+    searchfilters: {
         BRAND:{type:string, values:{}[] } ,
         COLOR: {type:string, values:{}[] },
         PRICE: {type:string, values:{}[] },
     };
-    products: {}[]
+    products: Product[]
+}
+
+interface SearchPageActions {
+    readCache(data);
+    updateFilters(data);
+    updateCatalog(data);
 }
 
 interface SearchPageState {
-    sortedProducts: {}[],
+    sortedProducts: Product[],
     sortby: string,
     minprice: number
     maxprice: number;
@@ -21,11 +32,7 @@ interface SearchPageState {
     colors: {}[];
 }
 
-const mapStateToProps = (state) => {
-
-}
-
-export default class SearchPage extends React.Component<SearchPageProps, SearchPageState> {
+class SearchPage extends React.Component<SearchPageProps & SearchPageActions, SearchPageState> {
     constructor(props) {
         super(props);
         this.state = {
@@ -36,6 +43,39 @@ export default class SearchPage extends React.Component<SearchPageProps, SearchP
             brand: '',
             colors: []
         }
+    }
+
+    componentDidMount = () => {
+        let CacheState = JSON.parse(
+            window.localStorage.getItem(CONST_STORAGE.SEARCH_STORAGE)
+        );
+        console.log(CONST_STORAGE.SEARCH_STORAGE, CacheState);
+        if (CacheState !== null && CacheState.filtersList !== undefined) {
+            this.props.readCache(CacheState);
+        } else {
+            fetch(CONST_URLS.SEARCH_FILTERS_URL).then(response => {
+                if (response.status !== 200) {
+                console.log("Could not load search filters!");
+                return;
+                }
+                response.json().then(data => {
+                    this.props.updateFilters(data)
+                });
+            });
+        }
+        fetch(CONST_URLS.SEARCH_PRODUCTS_URL).then(response => {
+            if (response.status !== 200) {
+                console.log("Could not fetch list of ");
+                return;
+            }
+            response.json().then(data => {
+                console.log("PRODUCTS FROM API", data);
+                this.props.updateCatalog(data);
+                this.setState({
+                    sortedProducts: [...this.props.products]
+                })
+            });
+        });
     }
 
     updateList = () => {
@@ -101,3 +141,39 @@ export default class SearchPage extends React.Component<SearchPageProps, SearchP
         </div>
     }
 }
+
+const mapStateToProps = state => {
+    return {
+        searchfilters: state.SearchAppReducer.filtersList,
+        products: state.SearchAppReducer.products
+    }
+}
+
+const mapDispatchToProps = dispatch => {
+    return {
+        readCache: (CacheState) => {
+            console.log(CacheState);
+            dispatch({
+                type: "READ_SEARCH_CACHE",
+                filterBrand: CacheState.filtersList.BRAND,
+                filterColor: CacheState.filtersList.COLOR,
+                filterPrice: CacheState.filtersList.PRICE,
+                products: CacheState.products === undefined ? [] : CacheState.products
+            });
+        },
+        updateFilters: (data) => {
+            dispatch({
+                type: "UPDATE_SEARCH_FILTERS",
+                filterdata: data
+            });
+        },
+        updateCatalog: (data) => {
+            dispatch({
+                type: "UPDATE_SEARCH_CATALOG",
+                products: data.products
+            });
+        }
+    }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(SearchPage);
